## Description
Added Database Size Scaling to TXStore Benchmark Tests.
The new benchmarks are being tracked in [TXStore Benchmarks & Optimizations](https://smartcontract-it.atlassian.net/wiki/spaces/Framework/pages/1488158874/TxStore+Benchmarks+Optimizations).
Ticket: https://smartcontract-it.atlassian.net/browse/BCFR-1218

## Resulting Optimizations

### Find Attempts Requiring Receipt Fetch
⚙️ Optimized Query + Added DB Indexes

🗃️ DB Size Factor | ⏳ Before | ⚡ After | 🚀 Improvement
-- | -- | -- | --
1x | 94,166 ns/op | 90,521 ns/op | 🟢 3.9% faster
10x | 265,812 ns/op | 174,333 ns/op | 🟢 34.4% faster
100x | 3,094,477 ns/op | 1,180,665 ns/op | 🟢 61.8% faster
1000x | 17,026,631 ns/op | 10,385,318 ns/op | 🟢 39.0% faster