@amit-momin 2025-04-28
> ## Resulting Optimizations
> ### Find Attempts Requiring Receipt Fetch
> ⚙️ Optimized Query + Added DB Indexes
> 
> 🗃️ DB Size Factor	⏳ Before	⚡ After	🚀 Improvement
> 1x	94,166 ns/op	90,521 ns/op	🟢 3.9% faster
> 10x	265,812 ns/op	174,333 ns/op	🟢 34.4% faster
> 100x	3,094,477 ns/op	1,180,665 ns/op	🟢 61.8% faster
> 1000x	17,026,631 ns/op	10,385,318 ns/op	🟢 39.0% faster

After reading through the tests, think it might be better to present the data by the DB size (tx count in the DB) instead of a factor. Seems like it could vary depending on the test. Like in the case of `Find Attempts Requiring Receipt Fetch`, we populate the DB with 7 other tx for every 1 tx that we should fetch.